refactor: move SerialMockRadio test double out of the shipping package
Mechanical move (PR 2 of 2, after #2129 pointed all tests at the
tests-local seam): git-mv src/rigplane/backends/icom7610/drivers/
serial_stub.py onto tests/serial_stub.py, replacing the transitional
re-export shim. Relative imports (from ....x) are rewritten 1:1 to
absolute rigplane.x paths; no other code change.

The module docstring now states explicitly that this is a non-shipping
test double so TX-safety sweeps have a citable exclusion rationale.
Result: grep -rn "\.set_ptt(" src/ (MOR-999/MOR-1015/MOR-1016 bypass
sweep) no longer contains the two permanent serial_stub false
positives — every remaining hit is a real production TX path.

Docs: the radio-state-pipeline-validation recipe imports the double
from tests/ (sys.path insert), and the legacy-state-writer inventory
row records the new location.
