+++ docs/internals/legacy-state-writer-inventory.md
@@ -46,7 +46,7 @@ Decision statuses:
 | `runtime/radio_initial_state.py` | `migrated` | Initial sweeps flow through runtime receive/observation paths and seed `StateStore` for supported fields. | Keep hardware-dependent validation manual where profiles require it. |
 | `runtime/radio_state_snapshot.py` | `executor_cache_keep` | Best-effort restore fallback may use `_last_*` caches but does not define consumer freshness. | No Web/rigctld delivery path may prefer this cache over `StateStore`. |
 | `core/_state_cache.py` and re-export `_state_cache.py` | `executor_cache_keep` plus `compatibility_only` | Retained for runtime timeout fallback and import compatibility. | Public `state_cache` exposure is compatibility-only; new delivery code must use snapshots/projections. |
-| `backends/icom7610/drivers/serial_stub.py` | `migrated` plus `compatibility_only` | MOR-437: the fake serial backend is now `ObservationPollable` — `SerialMockRadio.create_observation_poller(...)` feeds the provider observation pipeline, so production reads no longer depend on `StateStore` bulk sync from the stub. The internal `RadioState`/`StateCache` mirrors remain only to drive the stub's deterministic CI-V responses and legacy `radio_state`/`state_cache` test accessors. | Keep the internal mirrors scoped to CI-V simulation/legacy accessors; new stub state must be observable via the observation poller, not consumed from the mutable mirrors. |
+| `tests/serial_stub.py` (moved out of `backends/icom7610/drivers/` — non-shipping test double) | `migrated` plus `compatibility_only` | MOR-437: the fake serial backend is now `ObservationPollable` — `SerialMockRadio.create_observation_poller(...)` feeds the provider observation pipeline, so production reads no longer depend on `StateStore` bulk sync from the stub. The internal `RadioState`/`StateCache` mirrors remain only to drive the stub's deterministic CI-V responses and legacy `radio_state`/`state_cache` test accessors. | Keep the internal mirrors scoped to CI-V simulation/legacy accessors; new stub state must be observable via the observation poller, not consumed from the mutable mirrors. |
 
 ## Web Runtime, Revisions, and Frontend Store
